+++ src/GameEngine/Utilities/Quaternion.h
@@ -14,13 +14,13 @@ class Quaternion
 		Quaternion( const Quaternion & q );
 
 		// Construct with specified real (s) and imaginary (v) components
-		Quaternion( float s, float vi, float vj, float vk );
+		Quaternion( double s, double vi, double vj, double vk );
 
 		// Construct with specified real (s) and imaginary (v) components
-		Quaternion( float s, const Vector& v );
+		Quaternion( double s, const Vector& v );
 
 		// Get reference to the specified component (0-based indexing: r, i, j, k)
-		float & operator[]( int index );
+		double & operator[]( int index );
 
 		// Addition
 		Quaternion operator+( const Quaternion& q ) const;
@@ -32,17 +32,17 @@ class Quaternion
 		Quaternion operator-( void ) const;
 
 		// Right scalar multiplication
-		Quaternion operator*( float c ) const;
+		Quaternion operator*( double c ) const;
 
 		// Scalar division
-		Quaternion operator/( float c ) const;
+		Quaternion operator/( double c ) const;
 
 		// Hamilton product
 		Quaternion operator*( const Quaternion& q ) const;
 
 	protected:
-		// scalar (float) part
-		float m_Scalar;
+		// scalar (double) part
+		double m_Scalar;
 
 		// vector (imaginary) part
 		Vector m_Vector;
